Move location of get_word_and_translation
