+++ voc_builder/openai_svc.py
@@ -9,6 +9,26 @@ from voc_builder.models import WordChoice, WordSample
 
 logger = logging.getLogger()
 
+
+def get_word_and_translation(text: str, known_words: Set[str]) -> WordSample:
+    """Get the most uncommon word in the given text, the result also include other
+    information such as meaning of the word and etc.
+
+    :param text: The text which needs to be translated
+    :param known_words: Words already known
+    :return: a `WordSample` object
+    :raise VocBuilderError: when unable to finish the API call or reply is malformed
+    """
+    try:
+        reply = query_openai(text, known_words)
+    except Exception as e:
+        raise VocBuilderError('Error querying OpenAI API: %s' % e)
+    try:
+        return parse_openai_reply(reply, text)
+    except ValueError as e:
+        raise VocBuilderError(e)
+
+
 # The prompt being used to make word
 prompt_tmpl = dedent(
     '''
